Causal base (#6)
* Start on causal-base

Add support for inserting seqeunces
- associng
- spinning
- weaving

Sketch out causal-base. The foundational structure for managing complicated nested causal collection. Like a database, but causal and not great at query or large amounts of data :P

* Fix cljs tests

* Sketch out some more ideas around transactions

* Fix insert

* Flatten maps and lists into cb

* Transact works!

* Better support multi byte chars

* cb to edn

* Add causal base tests

* Add base tests to cljs runner

* Fix cljs

* Cleanup

* Add CausalBase type

* Improve causal-base public api

* Check off some of the roadmap

* Remove experimental fn

* Fix map/get to handle hidden nodes

* Sketch out undo function

* tmp cleanup

* Apply list weave logic to maps

This enables id based causes in maps instead of all causes needing to be keys. This adds flexibility to how nodes are hidden or shown in a map, which is required for infinite undo/redo.

Refactor this weave-node logic into the shared ns

* Early version of undo

might still change the special keys used so undo nodes can be distinguished from delete nodes

* Add redo to causal-base

* Add more cb tests

* Add history.hide/show types

* Add tests for shared nested site-ids

* Expose undo / redo

* Add examples to readme

* Add get-nodes and set-site-id to cb